TD-2116 Implements charting of data and moves some charting code to common TS
