+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Centre/Reports/Index.cshtml
@@ -8,7 +8,7 @@
   ViewData["Title"] = "Course reports";
 }
 
-<link rel="stylesheet" href="@Url.Content("~/css/trackingSystem/reports.css")" asp-append-version="true">
+<link rel="stylesheet" href="@Url.Content("~/css/shared/reports.css")" asp-append-version="true">
 
 <div class="nhsuk-grid-row">
   <div class="nhsuk-grid-column-one-quarter">
